raspitron/requirements.txt
requests
gTTS
pyalsaaudio
google-genai
numpy
scipy
